+++ src/components/sidebar/routePathView/NewRoutePathView.tsx
@@ -14,7 +14,7 @@ import { ToolbarStore } from '~/stores/toolbarStore';
 import RoutePathFactory from '~/factories/routePathFactory';
 import ViewHeader from '../ViewHeader';
 import RoutePathViewForm from './RoutePathViewForm';
-import * as s from './routePathViewTab.scss';
+import * as s from './routePathTab.scss';
 
 interface INewRoutePathViewProps {
     routeStore?: RouteStore;
@@ -100,7 +100,7 @@ class NewRoutePathView extends React.Component<INewRoutePathViewProps, INewRoute
     public render(): any {
         const currentRoutePath = this.state.currentRoutePath;
         return (
-        <div className={classnames(s.routePathViewTab, s.form)}>
+        <div className={classnames(s.routePathTab, s.form)}>
             <div className={s.formSection}>
                 <ViewHeader
                     header='Luo uusi reitinsuunta'
