+++ src/components/sidebar/routePathView/RoutePathView.tsx
@@ -5,9 +5,9 @@ import { match } from 'react-router';
 import { IRoutePath } from '~/models';
 import Loader from '~/components/shared/loader/Loader';
 import RoutePathService from '~/services/routePathService';
-import RoutePathViewTab from './RoutePathViewTab';
-import LinkNodeListViewTab from './LinkNodeListViewTab';
-import RoutePathViewTabButtons from './RoutePathViewTabButtons';
+import RoutePathTab from './RoutePathTab';
+import RoutePathLinksTab from './RoutePathLinksTab';
+import RoutePathTabButtons from './RoutePathTabButtons';
 import * as s from './routePathView.scss';
 
 interface IRoutePathViewState {
@@ -56,14 +56,14 @@ class RoutePathView extends React.Component<IRoutePathViewProps, IRoutePathViewS
         switch (this.state.selectedTabIndex) {
         case 0: {
             return (
-                <RoutePathViewTab
+                <RoutePathTab
                     routePath={this.state.routePath!}
                 />
             );
         }
         case 1: {
             return (
-                <LinkNodeListViewTab
+                <RoutePathLinksTab
                     routePath={this.state.routePath!}
                 />
             );
@@ -91,7 +91,7 @@ class RoutePathView extends React.Component<IRoutePathViewProps, IRoutePathViewS
         return (
             <div className={s.routePathView}>
                 <div className={s.flexInnerRow}>
-                    <RoutePathViewTabButtons
+                    <RoutePathTabButtons
                         tabs={routePathViewTabs}
                         selectedTab={routePathViewTabs[this.state.selectedTabIndex]}
                         selectTab={this.selectTab}
