+++ src/components/sidebar/routePathView/routePathTab.scss
@@ -2,7 +2,7 @@
 @import '../../shared/styles/flexbox';
 @import '../../shared/styles/form';
 
-.routePathViewTab {
+.routePathTab {
     padding: 15px 15px 45px 15px;
 
     .routeInformationContainer {
