+++ src/components/sidebar/routePathView/routePathTabButtons.scss
@@ -1,6 +1,6 @@
 @import '../../shared/styles/common';
 
-.routePathViewTabButtonsView {
+.routePathTabButtonsView {
     width:100%;
     height:40px;
     background-color: $lightGrey;
